Allow to poll.publish in started state, handle stop activities. Add tests.
